+++ src/main/java/com/tagmycode/sdk/exception/TagMyCodeUnauthorizedException.java
@@ -5,4 +5,8 @@ public class TagMyCodeUnauthorizedException extends TagMyCodeException {
     public TagMyCodeUnauthorizedException() {
         super("Unauthorized");
     }
+
+    public TagMyCodeUnauthorizedException(String message) {
+        super(message);
+    }
 }