+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## UNRELEASED
+### Added
+- throws TagMyCodeUnauthorizedException if refresh token is not valid
+
 ## [1.2.0] - (2017-11-05)
 ### Changed
 - renamed LanguagesCollection and SnippetsCollection